:sparkles: #130 - evalatuion create schedular(resolve conflict2)
