--- src/main/kotlin/com/teamsparta/tikitaka/domain/evaluation/service/EvaluationService.kt
@@ -1,7 +1,0 @@
-package com.teamsparta.tikitaka.domain.evaluation.service
-
-import com.teamsparta.tikitaka.domain.match.model.SuccessMatch
-
-interface EvaluationService {
-    fun createEvaluationsForMatch(match: SuccessMatch)
-}
