+++ src/main/kotlin/com/teamsparta/tikitaka/domain/evaluation/service/v3/EvaluationScheduler.kt
@@ -1,4 +1,4 @@
-package com.teamsparta.tikitaka.domain.evaluation.service
+package com.teamsparta.tikitaka.domain.evaluation.service.v3
 
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Service
