+++ src/main/kotlin/com/teamsparta/tikitaka/domain/evaluation/service/v3/EvaluationService.kt
@@ -2,9 +2,11 @@ package com.teamsparta.tikitaka.domain.evaluation.service.v3
 
 import com.teamsparta.tikitaka.domain.evaluation.dto.EvaluationRequest
 import com.teamsparta.tikitaka.domain.evaluation.dto.EvaluationResponse
+import com.teamsparta.tikitaka.domain.match.model.SuccessMatch
 import com.teamsparta.tikitaka.infra.security.UserPrincipal
 
 interface EvaluationService {
     fun evaluate(evaluationId: Long, principal: UserPrincipal, request: EvaluationRequest): EvaluationResponse
     fun calculateAndUpdateScores()
+    fun createEvaluationsForMatch(match: SuccessMatch)
 }
