+++ src/main/kotlin/com/teamsparta/tikitaka/domain/evaluation/service/v3/EvaluationServiceImpl.kt
@@ -4,7 +4,9 @@ import com.teamsparta.tikitaka.domain.common.exception.AccessDeniedException
 import com.teamsparta.tikitaka.domain.common.exception.ModelNotFoundException
 import com.teamsparta.tikitaka.domain.evaluation.dto.EvaluationRequest
 import com.teamsparta.tikitaka.domain.evaluation.dto.EvaluationResponse
+import com.teamsparta.tikitaka.domain.evaluation.model.Evaluation
 import com.teamsparta.tikitaka.domain.evaluation.repository.EvaluationRepository
+import com.teamsparta.tikitaka.domain.match.model.SuccessMatch
 import com.teamsparta.tikitaka.domain.team.repository.TeamRepository
 import com.teamsparta.tikitaka.domain.users.repository.UsersRepository
 import com.teamsparta.tikitaka.infra.security.UserPrincipal
@@ -18,6 +20,7 @@ class EvaluationServiceImpl(
     private val evaluationRepository: EvaluationRepository,
     private val usersRepository: UsersRepository,
     private val teamRepository: TeamRepository,
+    private val successMatchRepository: SuccessMatchRepository,
 ) : EvaluationService {
 
     @Transactional
@@ -64,5 +67,29 @@ class EvaluationServiceImpl(
 
             teamRepository.save(team)
         }
+    }
+
+    @Transactional
+    override fun createEvaluationsForMatch(match: SuccessMatch) {
+        val hostTeamEvaluation = Evaluation(
+            evaluatorTeamId = match.hostTeamId,
+            evaluateeTeamId = match.guestTeamId,
+            evaluatorId = match.guestId,
+            createdAt = LocalDateTime.now(),
+        )
+        val guestTeamEvaluation = Evaluation(
+            evaluatorTeamId = match.guestTeamId,
+            evaluateeTeamId = match.hostTeamId,
+            evaluatorId = match.hostId,
+            createdAt = LocalDateTime.now(),
+        )
+
+        evaluationRepository.save(hostTeamEvaluation)
+        evaluationRepository.save(guestTeamEvaluation)
+
+        match.evaluationCreatedTrue()
+        successMatchRepository.save(match)
+
+
     }
 }
